requirements.txt
requests>=2.31
PyYAML>=6.0
beautifulsoup4>=4.12
playwright>=1.41
Flask>=3.0
